44. Wildcard Matching
My solution is the DFS function but it is time limit exceeded...
I am so sad and  confused. I feel the thinking of DFS function is looked like the "while" function.
They are using the same thinking that finding the first character follow '*', and start matching string s and p.

At the last, I understand why DFS is too slow: it search every * and then recall when it is false, but it is not necessary.
Attention: when meet a new '*' character, we can ignore previous '*'. Because the rule 2: '*' can matching any string.